Prevent updates of actions in deprecated workflows
Wo don't want to be bothered with reviews of update PRs for actions in
the workflows that have been deprecated. However, there will probably
some non-reusable workflows for the actions that we'll be added to this
repository in the future, hence the Dependabot updates should not be
switched off entirely. Where possible, the versions have been changed
to rolling major version tags that are updated in the background. Some
actions that don't offer rolling tags have been excluded from updates,
but those are actually actions we'd probably not use in this repository
again.